feat(database): 添加数据库表结构定义和BCrypt测试类
- 创建了完整的PostgreSQL数据库表结构，包括用户、渠道、模型、订阅等核心表
- 添加了序列用于主键自增，确保数据唯一性和一致性
- 定义了权限控制、日志记录、配额管理等关键功能的数据表
- 创建BCryptTest测试类，实现密码加密和验证功能
- 配置了表间关联关系和索引以优化查询性能

+++ src/main/java/cc/flexapi/model/dto/AbilityDto.java
@@ -24,7 +24,7 @@ public class AbilityDto {
     /**
      * channel_id
      */
-    private Integer channelId;
+    private Long channelId;
 
     /**
      * enabled
@@ -39,7 +39,7 @@ public class AbilityDto {
     /**
      * weight
      */
-    private Integer weight;
+    private Long weight;
 
     /**
      * tag
