+++ src/main/java/cc/flexapi/model/dto/CheckinDto.java
@@ -14,12 +14,12 @@ public class CheckinDto {
     /**
      * id
      */
-    private Integer id;
+    private Long id;
 
     /**
      * user_id
      */
-    private Integer userId;
+    private Long userId;
 
     /**
      * 格式: YYYY-MM-DD
@@ -29,7 +29,7 @@ public class CheckinDto {
     /**
      * quota_awarded
      */
-    private Integer quotaAwarded;
+    private Long quotaAwarded;
 
     /**
      * created_at
