+++ src/main/java/cc/flexapi/model/dto/QuotaDataDto.java
@@ -14,12 +14,12 @@ public class QuotaDataDto {
     /**
      * id
      */
-    private Integer id;
+    private Long id;
 
     /**
      * user_id
      */
-    private Integer userId;
+    private Long userId;
 
     /**
      * username
@@ -39,16 +39,16 @@ public class QuotaDataDto {
     /**
      * token_used
      */
-    private Integer tokenUsed;
+    private Long tokenUsed;
 
     /**
      * count
      */
-    private Integer count;
+    private Long count;
 
     /**
      * quota
      */
-    private Integer quota;
+    private Long quota;
 
 }
