+++ src/main/java/cc/flexapi/model/dto/SubscriptionPreConsumeRecordDto.java
@@ -14,7 +14,7 @@ public class SubscriptionPreConsumeRecordDto {
     /**
      * id
      */
-    private Integer id;
+    private Long id;
 
     /**
      * request_id
@@ -24,12 +24,12 @@ public class SubscriptionPreConsumeRecordDto {
     /**
      * user_id
      */
-    private Integer userId;
+    private Long userId;
 
     /**
      * user_subscription_id
      */
-    private Integer userSubscriptionId;
+    private Long userSubscriptionId;
 
     /**
      * pre_consumed
