+++ src/main/java/cc/flexapi/model/dto/UserSubscriptionDto.java
@@ -14,17 +14,17 @@ public class UserSubscriptionDto {
     /**
      * id
      */
-    private Integer id;
+    private Long id;
 
     /**
      * user_id
      */
-    private Integer userId;
+    private Long userId;
 
     /**
      * plan_id
      */
-    private Integer planId;
+    private Long planId;
 
     /**
      * amount_total
