+++ src/main/java/cc/flexapi/model/po/CustomOAuthProviderPo.java
@@ -21,7 +21,7 @@ public class CustomOAuthProviderPo {
      */
     @Id
     @Column("id")
-    private Integer id;
+    private Long id;
 
     /**
      * Display name, e.g., "GitHub Enterprise"
@@ -117,7 +117,7 @@ public class CustomOAuthProviderPo {
      * 0=auto, 1=params, 2=header (Basic Auth)
      */
     @Column("auth_style")
-    private Integer authStyle;
+    private Long authStyle;
 
     /**
      * JSON policy for access control based on user info
