+++ src/main/java/cc/flexapi/model/po/SubscriptionPreConsumeRecordPo.java
@@ -20,7 +20,7 @@ public class SubscriptionPreConsumeRecordPo {
      */
     @Id
     @Column("id")
-    private Integer id;
+    private Long id;
 
     /**
      * request_id
@@ -32,13 +32,13 @@ public class SubscriptionPreConsumeRecordPo {
      * user_id
      */
     @Column("user_id")
-    private Integer userId;
+    private Long userId;
 
     /**
      * user_subscription_id
      */
     @Column("user_subscription_id")
-    private Integer userSubscriptionId;
+    private Long userSubscriptionId;
 
     /**
      * pre_consumed
