+++ src/main/java/cc/flexapi/model/po/TokenPo.java
@@ -21,13 +21,13 @@ public class TokenPo {
      */
     @Id
     @Column("id")
-    private Integer id;
+    private Long id;
 
     /**
      * user_id
      */
     @Column("user_id")
-    private Integer userId;
+    private Long userId;
 
     /**
      * key
@@ -69,7 +69,7 @@ public class TokenPo {
      * remain_quota
      */
     @Column("remain_quota")
-    private Integer remainQuota;
+    private Long remainQuota;
 
     /**
      * unlimited_quota
@@ -99,7 +99,7 @@ public class TokenPo {
      * used quota
      */
     @Column("used_quota")
-    private Integer usedQuota;
+    private Long usedQuota;
 
     /**
      * group
