+++ src/main/java/cc/flexapi/model/po/TwoFAPo.java
@@ -21,13 +21,13 @@ public class TwoFAPo {
      */
     @Id
     @Column("id")
-    private Integer id;
+    private Long id;
 
     /**
      * user_id
      */
     @Column("user_id")
-    private Integer userId;
+    private Long userId;
 
     /**
      * TOTP密钥，不返回给前端
@@ -45,7 +45,7 @@ public class TwoFAPo {
      * failed_attempts
      */
     @Column("failed_attempts")
-    private Integer failedAttempts;
+    private Long failedAttempts;
 
     /**
      * locked_until
