+++ src/main/java/cc/flexapi/model/po/UserSubscriptionPo.java
@@ -20,19 +20,19 @@ public class UserSubscriptionPo {
      */
     @Id
     @Column("id")
-    private Integer id;
+    private Long id;
 
     /**
      * user_id
      */
     @Column("user_id")
-    private Integer userId;
+    private Long userId;
 
     /**
      * plan_id
      */
     @Column("plan_id")
-    private Integer planId;
+    private Long planId;
 
     /**
      * amount_total
