+++ src/main/java/cc/flexapi/model/request/UserManageRequest.java
@@ -6,12 +6,11 @@ import lombok.Data;
 
 
 @Data
-
 public class UserManageRequest {
 
 
     @NotBlank(message = "用户id不能为空")
-    private Integer id;
+    private Long id;
 
 
     @NotBlank(message = "操作不能为空")
