+++ src/main/java/cc/flexapi/model/vo/CheckinVo.java
@@ -15,13 +15,13 @@ public class CheckinVo {
     /**
      * id
      */
-    private Integer id;
+    private Long id;
 
     /**
      * user_id
      */
     @JsonProperty("user_id")
-    private Integer userId;
+    private Long userId;
 
     /**
      * 格式: YYYY-MM-DD
@@ -33,7 +33,7 @@ public class CheckinVo {
      * quota_awarded
      */
     @JsonProperty("quota_awarded")
-    private Integer quotaAwarded;
+    private Long quotaAwarded;
 
     /**
      * created_at
