+++ src/main/java/cc/flexapi/model/vo/PasskeyCredentialVo.java
@@ -16,13 +16,13 @@ public class PasskeyCredentialVo {
     /**
      * id
      */
-    private Integer id;
+    private Long id;
 
     /**
      * user_id
      */
     @JsonProperty("user_id")
-    private Integer userId;
+    private Long userId;
 
     /**
      * base64 encoded
@@ -45,7 +45,7 @@ public class PasskeyCredentialVo {
      * sign_count
      */
     @JsonProperty("sign_count")
-    private Integer signCount;
+    private Long signCount;
 
     /**
      * clone_warning
